switch logging to proper usage
The logging methods do string interpolation themselves

The first arg to logging.debug() is formally defined to be a message
format string.

From the reference:

https://docs.python.org/2/library/logging.html#logging.Logger.debug

  Logger.debug(msg, *args, **kwargs)

Logs a message with level DEBUG on this logger. The msg is the message
format string, and the args are the arguments which are merged into msg
using the string formatting operator. (Note that this means that you can
use keywords in the format string, together with a single dictionary
argument.)

There are two keyword arguments in kwargs which are inspected: exc_info
which, if it does not evaluate as false, causes exception information to
be added to the logging message. If an exception tuple (in the format
returned by sys.exc_info()) is provided, it is used; otherwise,
sys.exc_info() is called to get the exception information

The reason logging does string interpolation it self is to implement
deferred interpolation.

String interpolation involves evaluating arguments, so it can introduce
significant computation.  The logging module tries to be smart about
deferring interpolation until the last possible moment.

The logging methods check isEnabledFor for the log level and won't
interpolate if the level is not enabled.

https://github.com/python/cpython/blob/2.7/Lib/logging/__init__.py#L1178

     def warning(self, msg, *args, **kwargs):
        if self.isEnabledFor(WARNING):
            self._log(WARNING, msg, args, **kwargs)

logging actually waits to interpolate the string in LogRecord.getMessage()

https://github.com/python/cpython/blob/2.7/Lib/logging/__init__.py#L328
        if self.args:
            msg = msg % self.args

Change-Id: I0090dcbc408200b6f2471748eae0c5a763da2e37
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>