Implements sha256 exception functionality
This patch adds functionality for file checksum verfications for
binary files.

The master_list.yaml binaries yaml directive now only contains
simple exceptions (for common artefacts that are gitignored)

Each project_exception file now has a filename and a sha256 hash.

If a binary file is not found, or the hash is mismatched,
it will output the hash for the user to include in an exception
patch.

This functionality has been added to complete project scans
and patchset scans

JIRA: RELENG-240

Change-Id: Iafa5710f4a0da192fc74335b1200b504413f8a8b
Signed-off-by: lhinds <lhinds@redhat.com>