Update patch set 1
Patch Set 1: Code-Review+2

Thanks for your contribution Dan!

Perhaps not for this patch, but it would be nice if we could use a sha256 file integrity patch check on the dovetail script. The potential risk is that if someone swapped out the img for a nefarious img, that image would get deployed throughout all of the POD and others who deploy opnfv.

I will raise this on the infra meeting today, and let's see what they say.

Patch-set: 1
Label: Code-Review=+2