Handle missing / deleted / renamed files correctly.
Previously git rm or mv'ed files would be listed in the patchset.

Anteater would then attempt to open the files and fail (as they don't
exist).

This patch resolves the issue by first not sys.exit'ing on a file not
existing, and also not logging with ERROR level (which made the
security audit job incorrectly fail).

JIRA: RELENG-302

Change-Id: I6a0c56f691f4d80aca1b3509472c1d6e56d492e1
Signed-off-by: lhinds <lhinds@redhat.com>