[daisy] Add wget related exceptions for Daisy
Violation this exceptions is trying to solve is wget plus variables.
But I think what I am doing is definitly wrong, because the value of
variables can change.

Sorry but I really do not know how to deal with violations like
these, so raise the question and tring to find answer. Thanks!

Change-Id: I5b44b15510e4cde8979b914f06e866366f0a4e4b
Signed-off-by: Zhijiang Hu <hu.zhijiang@zte.com.cn>